==> NBS/01A_read_file.ipynb <==
# Read the file with online food delivery data
by Smahi

## Scope
- Read the csv file in pandas dataframe.
- Print head of the dataframe.

## Dataset info
So the dataset contains information like:

* the age of the customer
* marital status of the customer
* occupation of the customer
* monthly income of the customer
- educational qualification of the customer
- family size of the customer
- latitude and longitude of the location of the customer
- pin code of the residence of the customer
- did the customer order again (Output)
- Feedback of the last order (Positive or Negative)

## Summary
- Successfully read the onlinefood.csv file in a pandas dataframe.
- The dataset has 388 rows and 13 columns.

## Imports

# Import libraries
import pandas as pd

# Read csv
df = pd.read_csv("C:/Users/SMAHI/Desktop/Online-food-delivery/Data/onlinefood.csv")

# Preview
df.head()

# Shape
df.shape
