==> NBS/02A_basic_stats_cleaning.ipynb <==
# Check for basic stats and clean the dataset
by Smahi

## Scope
- Check for null values in every column.
- Check for basic stats for the numeric columns.
- Filter the unnamed column if it is irrelevant.
- Save the processed data in a new csv

## Summary
- Changed the name of the columns in more easy to use snake_case.
- Droped the Unnamed:12 column as it was the replica of Output column.

## Imports

# Import libraries
import pandas as pd

# Read csv
df = pd.read_csv("C:/Users/SMAHI/Desktop/Online-food-delivery/Data/onlinefood.csv")

# Preview
df.head()

# Shape
df.shape

# Check for col names
df.columns

# Check for null values
df.isnull().sum()

There are no null values

# Check the data type for all columns
df.dtypes

# Use describe for basic stats
df.describe()

- Renaming column names to snake_case format

# Rename the columns in snake_format
df.rename(columns={'Marital Status': 'Marital_status',
                   'Monthly Income': 'Monthly_income',
                   'Educational Qualifications': 'Education',
                   'Family size':'Family_size',
                  'Pin code':'Pin_code'}, inplace=True)

# Preview
df.tail()


df['Unnamed: 12'].value_counts()

df['Output'].value_counts()

# Check if 'Column1' is a replica of 'Column2'
is_replica = (df['Unnamed: 12'] == df['Output']).all()

is_replica

# Drop column Unnamed:12
df.drop('Unnamed: 12',axis=1,inplace=True)

# Preview
df.head()

- It can be noticed that there are multiple categorical columns in the dataset. So printing those categories

# Print categories with count
df.Gender.value_counts()

# Print categories with count
df.Marital_status.value_counts()

# Print categories with count
df.Occupation.value_counts()

# Print categories with count
df.Monthly_income.value_counts()

# Print categories with count
df.Education.value_counts()

# Print categories with count
df.Family_size.value_counts()

# Print categories with count
df.Feedback.value_counts()

# Print categories with count
df.Output.value_counts()

# Save the DataFrame to a new CSV file
new_csv_path = r'C:/Users/SMAHI/Desktop/Online-food-delivery/Data/clean_data.csv'
df.to_csv(new_csv_path, index=False)

# Confirm that the DataFrame has been saved
print(f"DataFrame has been saved to {new_csv_path}")
