==> NBS/06A_correlation&relation_analysis.ipynb <==
# Correlation and Relationship analysis
by Smahi

## Scope
Correlations and Relationships:
- a. Is there a correlation between age and monthly income?
- b. Do marital status and family size have any relationship?
- c. Is there a relationship between educational qualifications and occupation?
- d. Are there correlations between the Output (online food orders) and any other variables?

## Summary
- There is a positive relationship between `Age` and `Monthly_income`.


- Marital status and its relationship with family size.

       Married:
        - Most individuals who are married have family sizes ranging from 2 to 6.
        - The most common family size for married individuals is 3, followed by 2 and 4.

       Prefer not to say:
        - Individuals who prefer not to disclose their marital status have smaller family sizes.
        - The most common family size for this group is 2.

       Single:
        - Singles have a diverse range of family sizes, but the most common family size is 3, followed by 2 and 4.
        - There are more single individuals with larger families (e.g., size 4, 5, 6) compared to the "Married" group.

- Occupation and its relationship with education

    **Occupation**
    
        Employee:
        - Most employees have a graduate-level education, with a substantial number having post-graduate qualifications as well.
        - There are no employees with only school or uneducated qualification.

        Housewife:
        - Housewives in this dataset have varying educational backgrounds.
        - The majority of housewives have a graduate or post-graduate education.

        Self-Employed:
        - Self-employed individuals have diverse educational backgrounds.
        - A significant number of self-employed individuals have graduate or post-graduate qualifications, but there are also some with school-level education.

        Student:
        - Students, as expected, are predominantly found in the "Student" occupation category.
        - Most students have either a graduate or post-graduate education.

    **Education**
    
        Graduate:
        - Graduates are spread across various occupations, with a notable presence in employee and student categories.

        Ph.D.:
        - Individuals with a Ph.D. are mainly employed as employees or students, with a small representation in the self-employed category.

        Post Graduate:
        - Post-graduates are found in all occupations, with a significant number in the student category.

        School:
        - Those with a school-level education are only found in the housewife and self-employed categories.

        Uneducated:
        - Individuals categorized as "Uneducated" are found in housewife and self-employed categories, with very small numbers

- The chi-square tests indicate that "Feedback," "Marital_status," and "Occupation" are strongly associated with online food orders ("Output"). 
- The "Education" variable also shows some evidence of association, but to a lesser extent. 
- The "Gender" variable, based on the p-value, does not seem to be significantly related to online food orders in this analysis

## Imports

import pandas as pd
from scipy.stats import chi2_contingency
import altair as alt
import seaborn as sns
import matplotlib.pyplot as plt

# Load the dataset
df = pd.read_csv('C:/Users/SMAHI/Desktop/Online-food-delivery/Data/clean_data.csv')

# Preview
df.head()

# Shape
df.shape

### a. Is there a correlation between age and monthly income?

# Different types of monthly_income
df.Monthly_income.unique()

# Convert Monthly_income to numerical values for correlation analysis
income_mapping = {'No Income': 0, 'Below Rs.10000': 1, 'More than 50000':2, '10001 to 25000':3, '25001 to 50000':4}
df['Monthly_income_numeric'] = df['Monthly_income'].map(income_mapping)


age_income_corr = df['Age'].corr(df['Monthly_income_numeric'])
print(f"Correlation between age and monthly income: {age_income_corr}")

The correlation coefficient between age and monthly income is 0.4559. This positive correlation suggests a moderate, positive relationship between age and monthly income in this dataset. 

### b. Do marital status and family size have any relationship?

marital_family_relation = pd.crosstab(df['Marital_status'], df['Family_size'])
print(f"Relationship between marital status and family size:\n{marital_family_relation}")

Married:
- Most individuals who are married have family sizes ranging from 2 to 6.
- The most common family size for married individuals is 3, followed by 2 and 4.

Prefer not to say:
- Individuals who prefer not to disclose their marital status have smaller family sizes.
- The most common family size for this group is 2.

Single:
- Singles have a diverse range of family sizes, but the most common family size is 3, followed by 2 and 4.
- There are more single individuals with larger families (e.g., size 4, 5, 6) compared to the "Married" group.

### c. Is there a relationship between educational qualifications and occupation?

edu_occ_relation = pd.crosstab(df['Education'], df['Occupation'])
print(f"Relationship between educational qualifications and occupation:\n{edu_occ_relation}")

**Occupation**:

Employee:
- Most employees have a graduate-level education, with a substantial number having post-graduate qualifications as well.
- There are no employees with only a school or uneducated in this dataset.

Housewife:
- Housewives in this dataset have varying educational backgrounds.
- The majority of housewives have a graduate or post-graduate education.

Self-Employed:
- Self-employed individuals have diverse educational backgrounds.
- A significant number of self-employed individuals have graduate or post-graduate qualifications, but there are also some with school-level education.

Student:
- Students, as expected, are predominantly found in the "Student" occupation category.
- Most students have either a graduate or post-graduate education.


**Education**:

Graduate:
- Graduates are spread across various occupations, with a notable presence in employee and student categories.

Ph.D.:
- Individuals with a Ph.D. are mainly employed as employees or students, with a small representation in the self-employed category.

Post Graduate:
- Post-graduates are found in all occupations, with a significant number in the student category.

School:
- Those with a school-level education are only found in the housewife and self-employed categories.

Uneducated:
- Individuals categorized as "Uneducated" are found in housewife and self-employed categories, with very small numbers

### d. Are there correlations between the Output (online food orders) and any other variables?

categorical_variables = ['Gender', 'Marital_status', 'Occupation', 'Education', 'Feedback']

output_corr = {}
for variable in categorical_variables:
    contingency_table = pd.crosstab(df['Output'], df[variable])
    chi2, p, _, _ = chi2_contingency(contingency_table)
    output_corr[variable] = p

output_corr = pd.Series(output_corr).sort_values()
print(f"Chi-square p-values for independence between Output and other categorical variables:\n{output_corr}")

**Null Hypothesis - Other categorical variables are independent to Output variable.**
- The chi-square tests indicate that "Feedback," "Marital_status," and "Occupation" are strongly associated with online food orders ("Output"). 
- The "Education" variable also shows some evidence of association, but to a lesser extent. 
- The "Gender" variable, based on the p-value, does not seem to be significantly related to online food orders in this analysis
